Re-organize accounts folder
